Delete test suite and move nondn role test. Otherwise it needs an additional start of the ldap server which slows down the tests (again).
