--- src/test/java/com/floragunn/dlic/auth/ldap/LdapAuthorizationTestNonLdapRoles.java
@@ -1,61 +1,0 @@
-package com.floragunn.dlic.auth.ldap;
-
-import org.elasticsearch.common.settings.Settings;
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-import com.floragunn.dlic.auth.ldap.backend.LDAPAuthorizationBackend;
-import com.floragunn.dlic.auth.ldap.srv.EmbeddedLDAPServer;
-import com.floragunn.dlic.auth.ldap.util.ConfigConstants;
-import com.floragunn.searchguard.user.User;
-
-public class LdapAuthorizationTestNonLdapRoles {
-    
-    protected EmbeddedLDAPServer ldapServer = null;
-
-    @Before
-    public final void startLDAPServer() throws Exception {
-        ldapServer = new EmbeddedLDAPServer();
-        ldapServer.start();
-        ldapServer.applyLdif("base.ldif");
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        if (ldapServer != null) {
-            ldapServer.stop();
-        }
-    }
-    
-    @Test
-    public void testLdapAuthorizationNonDNRoles() throws Exception {
-
-        final Settings settings = Settings.builder()
-                .putList(ConfigConstants.LDAP_HOSTS, "localhost:" + EmbeddedLDAPServer.ldapPort)
-                .put(ConfigConstants.LDAP_AUTHC_USERSEARCH, "(uid={0})")
-                .put(ConfigConstants.LDAP_AUTHC_USERBASE, "ou=people,o=TEST")
-                .put(ConfigConstants.LDAP_AUTHZ_ROLEBASE, "ou=groups,o=TEST")
-                .put(ConfigConstants.LDAP_AUTHZ_ROLENAME, "cn")
-                .put(ConfigConstants.LDAP_AUTHZ_RESOLVE_NESTED_ROLES, true)
-                .put(ConfigConstants.LDAP_AUTHZ_ROLESEARCH, "(uniqueMember={0})")
-                .put(ConfigConstants.LDAP_AUTHZ_USERROLENAME, "description, ou") // no memberOf OID
-                .put(ConfigConstants.LDAP_AUTHZ_ROLESEARCH_ENABLED, true)
-                .build();
-
-        final User user = new User("nondnroles");
-
-        new LDAPAuthorizationBackend(settings, null).fillRoles(user, null);
-
-        Assert.assertNotNull(user);
-        Assert.assertEquals("nondnroles", user.getName());
-        Assert.assertEquals(5, user.getRoles().size());
-        Assert.assertTrue("Roles do not contain non-LDAP role 'kibanauser'", user.getRoles().contains("kibanauser"));
-        Assert.assertTrue("Roles do not contain non-LDAP role 'humanresources'", user.getRoles().contains("humanresources"));
-        Assert.assertTrue("Roles do not contain LDAP role 'dummyempty'", user.getRoles().contains("dummyempty"));
-        Assert.assertTrue("Roles do not contain non-LDAP role 'role2'", user.getRoles().contains("role2"));
-        Assert.assertTrue("Roles do not contain non-LDAP role 'anotherrole' from second role name", user.getRoles().contains("anotherrole"));
-    }
-	
-}
